#19 pages/search 코드 정리
