+++ src/components/search/KindStudyComponent.tsx
